+++ packages/engine/Source/Scene/Geometry3DTileContent.js
@@ -1,7 +1,6 @@
 import Cartesian3 from "../Core/Cartesian3.js";
 import defaultValue from "../Core/defaultValue.js";
 import defined from "../Core/defined.js";
-import deprecationWarning from "../Core/deprecationWarning.js";
 import destroyObject from "../Core/destroyObject.js";
 import DeveloperError from "../Core/DeveloperError.js";
 import getJsonFromTypedArray from "../Core/getJsonFromTypedArray.js";
@@ -44,12 +43,6 @@ function Geometry3DTileContent(
   this._group = undefined;
 
   this._ready = false;
-
-  // This is for backwards compatibility. It can be removed once readyPromise is removed.
-  this._resolveContent = undefined;
-  this._readyPromise = new Promise((resolve) => {
-    this._resolveContent = resolve;
-  });
 
   initialize(this, arrayBuffer, byteOffset);
 }
@@ -117,26 +110,6 @@ Object.defineProperties(Geometry3DTileContent.prototype, {
   ready: {
     get: function () {
       return this._ready;
-    },
-  },
-
-  /**
-   * Gets the promise that will be resolved when the tile's content is ready to render.
-   *
-   * @memberof Geometry3DTileContent.prototype
-   *
-   * @type {Promise<Geometry3DTileContent>}
-   * @readonly
-   * @deprecated
-   * @private
-   */
-  readyPromise: {
-    get: function () {
-      deprecationWarning(
-        "Geometry3DTileContent.readyPromise",
-        "Geometry3DTileContent.readyPromise was deprecated in CesiumJS 1.104. It will be removed in 1.107. Wait for Geometry3DTileContent.ready to return true instead."
-      );
-      return this._readyPromise;
     },
   },
 
@@ -329,7 +302,6 @@ function initialize(content, arrayBuffer, byteOffset) {
 
   if (byteLength === 0) {
     content._ready = true;
-    content._resolveContent(content);
     return;
   }
 
@@ -550,7 +522,6 @@ Geometry3DTileContent.prototype.update = function (tileset, frameState) {
   if (defined(this._batchTable) && this._geometries.ready) {
     this._batchTable.update(tileset, frameState);
     this._ready = true;
-    this._resolveContent(this);
   }
 };